refactor: timetable adapter moved into adapters package, view init with adapter
